M3-293 Refactor : Permission 이름 수정
