+++ src/main/java/com/m3pro/groundflip/controller/PermissionController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.m3pro.groundflip.domain.dto.Response;
-import com.m3pro.groundflip.domain.dto.preference.PreferenceRequest;
-import com.m3pro.groundflip.service.PreferenceService;
+import com.m3pro.groundflip.domain.dto.permission.PermissionRequest;
+import com.m3pro.groundflip.service.PermissionService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -22,19 +22,19 @@ import lombok.extern.slf4j.Slf4j;
 @Tag(name = "permission", description = "권한 동의 API")
 @SecurityRequirement(name = "Authorization")
 public class PermissionController {
-	private final PreferenceService preferenceService;
+	private final PermissionService permissionService;
 
 	@Operation(summary = "서비스 알림 권한 동의", description = "서비스 푸시 알림을 받을지 받지 않을지 선택하는 api 이다.")
 	@PutMapping("/service-notification")
-	public Response<?> updateServiceNotification(@RequestBody PreferenceRequest preferenceRequest) {
-		preferenceService.updateServiceNotificationsPreference(preferenceRequest);
+	public Response<?> updateServiceNotification(@RequestBody PermissionRequest permissionRequest) {
+		permissionService.updateServiceNotificationsPreference(permissionRequest);
 		return Response.createSuccessWithNoData();
 	}
 
 	@Operation(summary = "마케팅 알림 권한 동의", description = "마케팅 푸시 알림을 받을지 받지 않을지 선택하는 api 이다.")
 	@PutMapping("/marketing-notification")
-	public Response<?> updateMarketingNotification(@RequestBody PreferenceRequest preferenceRequest) {
-		preferenceService.updateMarketingNotificationsPreference(preferenceRequest);
+	public Response<?> updateMarketingNotification(@RequestBody PermissionRequest permissionRequest) {
+		permissionService.updateMarketingNotificationsPreference(permissionRequest);
 		return Response.createSuccessWithNoData();
 	}
 }
