+++ src/main/java/com/m3pro/groundflip/domain/dto/permission/PermissionRequest.java
@@ -1,4 +1,4 @@
-package com.m3pro.groundflip.domain.dto.preference;
+package com.m3pro.groundflip.domain.dto.permission;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Data
 @Builder
 @Schema(title = "권한 동의 Body")
-public class PreferenceRequest {
+public class PermissionRequest {
 	@Schema(description = "사용자 ID", example = "5")
 	private Long userId;
